third_party/WebKit/LayoutTests/media/video-controls-show-on-focus-expected.txt

Test that video controls re-appear when the media (<video>) element is focused.

EXPECTED (getComputedStyle(controls).opacity == '0') OK
EXPECTED (getComputedStyle(controls).opacity == '1') OK
EXPECTED (getComputedStyle(controls).opacity == '1') OK
END OF TEST
